Added mesh file writing
When obj assets are loaded and processed into mesh classes, they are encoded to a file and saved with the '.mesh' extension.

This should make it quicker to load pre-processed obj files in the future, as the mesh files will not have to go through the lengthy indexing process.

Mesh files contain 3 major components:
 - Comments, containing useful readable data such as the time of creation.
 - Vertices, containing data for vertices.
 - Indices, containing data for indices.